requirements.txt
discord.py>=2.3.2
openai>=1.14.3
python-dotenv>=1.0.1
